pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-to-doc"
version = "0.1.0"
description = "Crawl videos from URLs and generate technical documentation with AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["video", "documentation", "AI", "crawler", "technical-writing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "yt-dlp>=2023.0.0",
    "openai>=1.0.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
whisper = [
    "openai-whisper>=20230314",
    "torch>=2.0.0",
]

[project.scripts]
video-to-doc = "video_to_doc.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/video-to-doc"
Issues = "https://github.com/yourusername/video-to-doc/issues"
